Move API types into lcdapi
