+++ packages/sdk38/src/index.ts
@@ -27,10 +27,10 @@ export {
   EncodeTxResponse,
   PostTxsResponse,
   NodeInfoResponse,
-  RestClient,
   SearchTxsResponse,
   TxsResponse,
-} from "./restclient";
+} from "./lcdapi";
+export { RestClient } from "./restclient";
 export { isMsgDelegate, isMsgSend, Msg, MsgDelegate, MsgSend } from "./msgs";
 export { Pen, Secp256k1Pen, makeCosmoshubPath } from "./pen";
 export { decodeBech32Pubkey, encodeBech32Pubkey, encodeSecp256k1Pubkey } from "./pubkey";
