+++ packages/sdk38/src/lcdclient.spec.ts
@@ -5,11 +5,11 @@ import { rawSecp256k1PubkeyToAddress } from "./address";
 import { Coin } from "./coins";
 import { isPostTxFailure } from "./cosmosclient";
 import { makeSignBytes } from "./encoding";
+import { TxsResponse } from "./lcdapi";
 import { LcdApiArray, LcdClient, normalizeArray } from "./lcdclient";
 import { parseLogs } from "./logs";
 import { MsgSend } from "./msgs";
 import { makeCosmoshubPath, Secp256k1Pen } from "./pen";
-import { TxsResponse } from "./restclient";
 import { SigningCosmosClient } from "./signingcosmosclient";
 import cosmoshub from "./testdata/cosmoshub.json";
 import {
