+++ packages/sdk38/types/index.d.ts
@@ -25,10 +25,10 @@ export {
   EncodeTxResponse,
   PostTxsResponse,
   NodeInfoResponse,
-  RestClient,
   SearchTxsResponse,
   TxsResponse,
-} from "./restclient";
+} from "./lcdapi";
+export { RestClient } from "./restclient";
 export { isMsgDelegate, isMsgSend, Msg, MsgDelegate, MsgSend } from "./msgs";
 export { Pen, Secp256k1Pen, makeCosmoshubPath } from "./pen";
 export { decodeBech32Pubkey, encodeBech32Pubkey, encodeSecp256k1Pubkey } from "./pubkey";
